requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
channels_redis==4.2.1
channels==4.2.0
charset-normalizer==3.4.1
colorama==0.4.6
Django==4.2.19
djangorestframework==3.15.2
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
msgpack==1.1.0
packaging==24.2
pip==25.0.1
pluggy==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
